Export log file per activation reference
This also moves logic for ADIF generation to plugins. Where no
activation is taking place, a single log is generate with any hunting
references present in that one file.

A QSO line is also generated per hunted reference, which will have no
effect on SOTA or WWFF as they are single activation only.