Add support for WAB Square in Operation Info
This requires additional package *proj4* for coordinate projection for
identifying WAB square from latitude/longitude of device location.

This introduces a new hook for `opSettings` for per operation
parameters. This felt like a better approach, as WAB isn't an activity
as such, just fact of your location, and is something that is "hunted",
and has no requirement to export/upload logs, etc.

Currently WAB square isn't export, but extension could be extended to
include adding to `COMMENT` field in ADIF (as is of interest to hunter,
not logger) or custom field.