Add Global Mountain Activity
Similar to SOTA, with summit to summit activity, but more flexible
reference and region in references, so less reference validation, and
no ability to add `-` or `/` automatically in S2S input.

GMA site does support `GMA_REF` and `SIG`/`SIG_INFO` with *GMA* as
`SIG` for logs. As `GMA_REF` isn't in ADIF standard, opted for
`SIG`/`SIG_INFO` as will ensure ADIF is valid for other applications.

This activity was based on SOTA with modifications due to similarity.