Add HamQTH lookup (#86)
* Add HamQTH lookup

This adds HamQTH lookup for call data. This is a free service but
requires login credentials.

This change includes a new account setting for entering log in details,
and new API for HamQTH, which is very similar to QRZ.com API.

* Improve handling of lookup merging

This ensures that when multiple lookups are done, that information is
merged in priority order, but also stops high priority empty data taking
preference of lower priority but present data.

This also resolved an issue with high priority data stopping lookup from
other services, when that data was empty string e.g. grid from free
QRZ.com lookup.

This does require that empty data for numbers is now cast as `null`
rather than `0`

* Fix listing sources on lookups

* Add HamQTH to release notes