Attempt to fetch multiple references for POTA spots
This utilises the known structure of the spots from fellow Ham2k PoLo
users to fetch their extra park references when hunting.

As this requires extra calls to POTA API, it's just too slow to do
when POTA returnging 100s of spots, so this added feature is only used
when selecting a spot.